requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
kafka-python==2.0.2
numpy==1.26.4
pandas==2.2.2
pendulum==3.0.0
py4j==0.10.9.7
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
time-machine==2.14.1
tzdata==2024.1
urllib3==2.2.1
